requirements.txt
aiogram>=3.4,<4
SQLAlchemy>=2.0,<3
python-dotenv>=1.0
flask>=3.0,<4
